refactor(src/components): change Products components
Refactor details, mobileButton and breadCrumb. Change types and props to.
