+++ src/components/ProductDetails/MobileButton.tsx
@@ -1,0 +1,45 @@
+import type { NextPage } from 'next';
+import { Button } from '../DefaultDesignComponents/Buttons';
+import { SmallParagraph } from '../DefaultDesignComponents/Typography';
+import Price from '../DefaultDesignComponents/Typography/Price';
+import { MobileButtonContainer } from './styles';
+
+interface PriceDataProps {
+  discount: number;
+  price: number;
+  priceMember: number;
+  priceNonMember: number;
+}
+
+interface MobileButtonProps {
+  priceData: PriceDataProps;
+}
+
+const MobileButton: NextPage<MobileButtonProps> = ({ priceData }) => {
+  const { discount, price, priceMember, priceNonMember } = priceData;
+  return (
+    <MobileButtonContainer>
+      <div className="mobile-price-info">
+        <SmallParagraph filled color="chenin" size="mini">
+          {discount}% OFF
+        </SmallParagraph>
+        <SmallParagraph
+          decoration="line-through"
+          size="xSmall"
+          type="grayLight"
+        >
+          {price.toFixed(2)}
+        </SmallParagraph>
+        <Price value={priceMember} size="small" />
+        <SmallParagraph size="mini" weight="bold" type="gray">
+          PREÇO NÃO SÓCIO R${priceNonMember.toFixed(2)}
+        </SmallParagraph>
+      </div>
+      <Button filled color="success" size="mobile">
+        Adicionar
+      </Button>
+    </MobileButtonContainer>
+  );
+};
+
+export default MobileButton;
